Refactored folder names. Created stockDB.go file
