--- Controllers/StockController.go
@@ -1,16 +1,0 @@
-package controllers
-
-import (
-	repositories "Main/Repositories"
-)
-
-type BaseHandler struct {
-	stockRepo repositories.StockRepository
-}
-
-// NewBaseHandler returns a new BaseHandler
-func NewBaseHandler(stockRepo repositories.StockRepository) *BaseHandler {
-	return &BaseHandler{
-		stockRepo: stockRepo,
-	}
-}
